Applied suggested fixes Created README.md Updated code to 3.6 and fixed running errors Moved persistence in another package Renamed project to respect vertx-examples conventions Added some comments
