+++ web-api-service-example/src/main/java/io/vertx/examples/webapiservice/WebApiServiceExampleMainVerticle.java
@@ -7,8 +7,8 @@ import io.vertx.core.eventbus.MessageConsumer;
 import io.vertx.core.http.HttpServer;
 import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.json.JsonObject;
+import io.vertx.examples.webapiservice.persistence.TransactionPersistence;
 import io.vertx.examples.webapiservice.services.TransactionsManagerService;
-import io.vertx.examples.webapiservice.services.impl.TransactionPersistence;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.api.contract.openapi3.OpenAPI3RouterFactory;
 import io.vertx.serviceproxy.ServiceBinder;
@@ -26,7 +26,7 @@ public class WebApiServiceExampleMainVerticle extends AbstractVerticle {
   private void startTransactionService() {
     serviceBinder = new ServiceBinder(vertx);
 
-    TransactionPersistence persistence = new TransactionPersistence();
+    TransactionPersistence persistence = TransactionPersistence.create();
 
     // Create an instance of TransactionManagerService and mount to event bus
     TransactionsManagerService transactionsManagerService = TransactionsManagerService.create(persistence);
@@ -41,7 +41,7 @@ public class WebApiServiceExampleMainVerticle extends AbstractVerticle {
    */
   private Future<Void> startHttpServer() {
     Future<Void> future = Future.future();
-    OpenAPI3RouterFactory.createRouterFactoryFromFile(this.vertx, getClass().getResource("/openapi.json").getFile(), openAPI3RouterFactoryAsyncResult -> {
+    OpenAPI3RouterFactory.create(this.vertx, "/openapi.json", openAPI3RouterFactoryAsyncResult -> {
       if (openAPI3RouterFactoryAsyncResult.succeeded()) {
         OpenAPI3RouterFactory routerFactory = openAPI3RouterFactoryAsyncResult.result();
 
@@ -51,8 +51,11 @@ public class WebApiServiceExampleMainVerticle extends AbstractVerticle {
         // Generate the router
         Router router = routerFactory.getRouter();
         server = vertx.createHttpServer(new HttpServerOptions().setPort(8080).setHost("localhost"));
-        server.requestHandler(router::accept).listen();
-        future.complete();
+        server.requestHandler(router).listen(ar -> {
+          // Error starting the HttpServer
+          if (ar.succeeded()) future.complete();
+          else future.fail(ar.cause());
+        });
       } else {
         // Something went wrong during router factory initialization
         future.fail(openAPI3RouterFactoryAsyncResult.cause());
