+++ web-api-service-example/src/main/java/io/vertx/examples/webapiservice/models/package-info.java
