+++ web-api-service-example/src/main/resources/openapi.json
